requirements.txt
langchain
llama-cloud-services
google-generativeai
streamlit
langchain-google-genai
langchain-community
python-dotenv
faiss-cpu
chromadb
pypdf
PyPDF2
transformers
sentence-transformers
numpy
pathlib
